draw out CD-CoT
